change double data type to BigDecimal for currency, modify db data types

+++ tests/java/com/allan/amca/junit/transaction/TransactionTest.java
@@ -2,6 +2,8 @@ package com.allan.amca.junit.transaction;
 
 import com.allan.amca.transaction.Transaction;
 
+import java.math.BigDecimal;
+
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
@@ -18,7 +20,7 @@ class TransactionTest {
     }
 
     protected void getAmount(final Transaction transaction,
-                             final double      expectedAmount) {
+                             final BigDecimal  expectedAmount) {
         assertThat(transaction.getTransactionAmount(), equalTo(expectedAmount));
     }
